requirements.txt
djangorestframework 
pillow 
python-dotenv 
google-generativeai 
google-auth-httplib2 
google-auth-oauthlib 
requests
asgiref==3.5.2
Django==4.1.3
mysqlclient
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6